+++ packages/compose/src/higher-order/with-state/index.js
@@ -6,7 +6,7 @@ import { Component } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import createHigherOrderComponent from '../create-higher-order-component';
+import createHigherOrderComponent from '../../utils/create-higher-order-component';
 
 /**
  * A Higher Order Component used to provide and manage internal component state