+++ packages/compose/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Master
+
+### New Features
+
+- Add the `useMediaQuery` and `useReducedMotion` hooks. 
+
 ## 3.0.0 (2018-11-15)
 
 ### Breaking Changes